add initial test for layer_list
